[compiler.h] Add a __RESTRICT macro
__restrict is a common C++ extension, which is legal in roughly the
same places as C's restrict keyword (that is, on declarations of
pointer variables, including parameters). While not especially useful
for C++'s own sake, this is useful for matching C interfaces which use
restrict. In particular, I want to use it in fdio when converting it
to C++, as fdio has many function signatures which parallel standard C
or POSIX ones containing restrict.

Test: compiles

Change-Id: I48498fbac4c00f9f5074ab8eab7b311bf03fed57