[examples] Add gain flag to WavRecorder, obey it in SignalGenerator
This CL adds a --gain flag to wav_recorder, so that the AudioIn gain
stage can be easily manipulated from an in-build example.

It also adds code in signal_generator to NOT call SetGain unless the
--gain flag has been specified (previously was always calling this).

Finally, it fixes a modulo bug in signal generator where the last
packet was not sent, if the number of samples aligned perfectly.

Test: build, CQ, manual testing using signal_generator, wav_recorder
(with --loopback) and audacity.

Change-Id: I83dde263fee5f3ba5fc471d627304faea5c2ac89