[netstack] remove integer conversions
This is slightly dangerous in that we're using unsigned types now which
are prone to silent underflow, but matches all the interfaces we consume
from the kernel.

While I'm here, change fifo{Read,Write} callers to check the status
before using the return value. AFAICT, the return value is undefined if
the status is anything other than ZX_OK.

TESTED=CQ

Change-Id: I80c00e4597ae792f98a2a66ff5b39001a7baeeb1