[mediaplayer] packet size is not buffer size
When PayloadBuffer was introduced, Packet depended on the size of its
associated PayloadBuffer to carry packet size information. Buffers
can be bigger than the payloads they contain, so the packet needs its
own notion of size. This CL addresses that.

TESTED: No behavior change, manual and CQ tests for regression
Change-Id: I6e8724001bde413cee5d0daaa448161ad7471c01