[fbl][containers] Fix a bug in SLL erase_next
Fix a bug in the implementation of the singly linked list's erase_next
method.  Specifically, the test to see if there was an element to
erase after the element specified by iter was incorrectly testing for
nullptr instead of testing for the sentinel value.

Update the unit tests to cover this case.  Also update the
documentation and API contract for erase_next in the doubly linked
list to match that of the singly linked list.  Specifically, it is an
error to attempt to erase the element after an an invalid iterator.

Tests: update fbl unit tests
Change-Id: I3edcadfb67b16be58f829f7031799129e31ff679