[legder] TryEvictPages can evict both completely synced, and empty pages.
There is no reason to only evict pages that are completely backed up,
and not those that are empty. On the other hand, it makes sense to
try to evict a page only if it is empty, e.g. because the page was
just closed and we want to evict it only if it won't change anything
from the user's prespective.

After this CL, TryEvictPages will check for both these conditions. If
either is true, the page can and is evicted.

Test: PageEvictionManagerTest.EvictUnsyncedButEmptyPages
      PageEvictionManagerTest.EvictSyncedAndNotEmptyPages
      PageEvictionManagerTest.DontEvictUnsyncedNotEmptyPages
Change-Id: I54d0e606c5b32065a6b78746238f31984be263bc