[optee][tee] Overhaul of OP-TEE messages
This CL makes the following changes to the message representation within
the OP-TEE driver:
  * RpcMessage is a new class that replaces UnmanagedMessage and
    features new safety checks on memory before parsing contents.
  * Different RPC commands now subclass from RpcMessage and should
    perform general parameter validation within these classes.[1]
  * Replace ManagedMessage with a new Message type that has move-only
    semantics to better represent intended ownership rules.
  * New documentation :)

[1] To prevent the change from being XXL (for reviewers' sakes), just
the RPC command for loading trusted applications has been ported over to
using a subclass of RpcMessage for its message parameter parsing. The
remaining implemented RPC commands will be ported over in a subsequent
change.

Test: Just representation has been changed but behavior is preserved.
Confirmed on astro.

Change-Id: I0145fbbf5e983b3a699d5a10ae5c7540d45110c9