[kernel][dev][intel-rng] fix buffer overflow bug
Noticed this while converting the file to cpp. The code to trim the
amount of memory to memcpy into the outgoing buffer would overflow by up
to 8 bytes per loop.

Never showed up before because the calling code for this either always
asks for a multiple of 8 or copies into a buffer that is at least a
multiple of 8.

Perhaps worth writing an in kernel unit test.

Change-Id: I4363f63360f121d444e4a6df9c8ead5de96b2a1b
Tested: build and verify that the copy does not overflow manually.