[wayland] Add types for the registry.
The registry allows some global interfaces to be registered for the
server. When a client connects, it will send a message asking the
server to announce it's global interfaces that can be bound. The client
will then bind instances of the interfaces it's interested in to a new
object ID that can be used to identify that instance on the wire.

Test: fx run-test wayland_bridge_bin_test_rustc
Change-Id: Ib9b01aa6a8e21bec45691d57d4e564d1aaf3855c