[kernel][lib][cbuf] fix potential deadlock
Signalling the cbuf signal inside the cbuf's spinlock is potentially bad
since it can reschedule while holding the lock. Move the event signal
outside of the lock.

This isn't happening now since all the callers of this routine are
currently inside an interrupt which implicitly disables local
rescheduling.

Change-Id: I3aca2963bfd7252817aa90d8dd60931ebb2a328e
Tested: compile and run, validating that serial IO continues to work.