[fdio] Merge pipe() and socketpair() impls
As far as I can tell, there's no reason for these implementations to be
separate. Merging the two means we'll be able to recover the client type
of a zx::socket from the information in the kernel. Specifically, a
stream socket with no control plane will map to a pipe (aka socketpair).

Test: pipe and socketpair tests continue to pass
Change-Id: If0251b662daa1ad4ce4017c7d949016204c667a1