[debugger] Format non-interactive terminal output.
The code switches between "raw" terminal mode (when editing a line) and
non-raw mode (when outputting stuff). Non-raw mode interprets LF a CRLF
sp each "\n" in a string goes to the beginning of the next line.
Currently the code expects the startup state to be non-raw mode and will
restore to that state for normal output.

If the terminal is corrupted (usually by a program running in raw mode
crashing or getting killed), the terminal will be in raw mode at startup
and all non-raw output will be corrupted.

This change forces on the post-processing for non-raw mode for correct
output in this case.

TEST=manual

Change-Id: I1aacc50c72ecc3c35d0f3c2f2276497a65b74906