[kernel] Make get_info return mappings for unstarted process
As pointed out in the previous documentation, ZX_ERR_BAD_STATE
would be returned if zx_object_get_info was called with
ZX_INFO_PROCESS_MAPS on a process that has not been started.
This does not appear to be behavior that is needed by anything.
For testing elf-search however it is very helpful to be able to
create a dummy process that has not been started and anylyze its
address space.

Test: CQ + manual
Change-Id: Id5a098409190dd4390ff1539bd483420dbdcef05