[media] Internal renderer/capturer rename
This CL is the first part of a multi-stage sequence that
changes the names of the primary audio FIDL interfaces,
AudioOut and AudioIn, to the more intuitive AudioRenderer
and AudioCapturer.

Specifically, this CL does not yet change any FIDL interfaces,
but migrates (within garnet/../media) all internal usage of the
terms audio_out, audio_in, AudioOut and AudioIn to renderer,
capturer, AudioRenderer and AudioCapturer (usually variables,
parameters or function names).

Test: build, CQ, signal_generator, wav_recorder

Change-Id: I912fd4eed29d7ff9e519781e178223d657f84a2d