[auth] Pass AuthenticationUIContext from device_shell to user_provider
Introduces a new method in device_shell fidl to support TokenManager
service and the following cl updates the device shell implementation in
the UI layer:
https://fuchsia-review.googlesource.com/c/topaz/+/200743

This cl also replaces a single tokenmanager handle used in user_runner to two
tokenmanager handles - one that is passed to the ledger and the other that is
used for agents. The rationale behind this change is clients shouldn't be minting
new token managers when AccountManager gets introduced, and also clients
shouldn't depend on auth provider configuration going forward.

TEST = CQ, Guest login and Manual login using both old and new interfaces (by
setting enable_token_manager=true and overriding account_provider_url to
"token_manager_rust")

Change-Id: I3743b84b1ad1d57e30dc990e216818bbee79b7a1