[debugger] Improve stepping, add logging.
Fixes a stepping bug where "next" would stop prematurely after stepping
out of a cross-module call. The problem was an "optimization" in the
step controller that returned early if single-stepping in an enpty
range. This bypassed the checking for contuing in the same line.
Removing the special case fixes the bug.

Added a thread controller logging system. Currently this is enabled with
an ifdef. Although I find these systems undesirable, stepping is
complicated and we are likely to have a neverending long tail of
complaints about stepping behavior. Logging the thread controllers makes
it much faster to diagnose these issues.

TEST=manual

Change-Id: I14131bd543db0d00da4de325ccf475dfd79e2f64