[astro][tee] Enable OP-TEE for Astro
In order to communicate with the OP-TEE OS on Astro, the Astro board
driver must add the OP-TEE device to the platform bus. This change
enables the OP-TEE device and allows it to map the Secure OS memory
region. This memory region is used for communication between the secure
world and the non-secure world.

Test: Manual testing on Astro to ensure device was initialized.
Change-Id: I6c2e6f05c9579a9999df1056ec3215a4f9ab9406