[cobalt-client]: Add CobaltLogger.
Added CobaltLoggerOptions to ease testing.
config_reader allows the client to decide how to construct the vmo
containing the configuration buffer.
system_connect could be exposed to the user, but essentially is used
to replace fdio_service_connect while testing, so we can use in process
services.

CobaltLogger does not make any assumption on Cobalt services
availability. This is achieved as follows:

* When attempting to send metrics to the service, we asynchronously
request a Logger. If this channel is closed or error occurs, we go
back to the beginning.

* If the request has been sent, and there is an existing channel
to thte LoggerFactory, then check if the reply has been written
into the channel. This is a wait of at most |CobaltOptions::deadline|.
Three options can happen on this step, he channel became readable,
timed out, or any other error. If it becomes readable, we look at the
response, if it times out, we dont log anything, and we will try again
on the next time we try to log. If we have any other error, we close the
channel and we connect to the factory once again.

* If the logger channel is open, then we synchronously Log the data(wait
for reply after sending the Log*Request). This allows us to know whether
the log was successfull or not.

* The logger provides the abilitty to wait for its logger_client to
become readable to allow deterministic testing, and race conditions
between writing to the channel and the data becoming available in the
dispatcher port.

TODO: Pipe the component data into EventBuffer and add when doing fidl
calls.

TESTS=cobalt-client-tests

Change-Id: I277ff178bfaec4c0f593770678c83bb13f792af3