[view_manager] Remove ViewToken and GetToken()
ViewOwner (the handles for each side of the channel) are used as the
unique tokens which can match a ViewOwner to its associated View.  In
particular, the handle<ViewOwner> is used by the client to identify the
view to the view_manager when linking it into the scene hierarchy with
addChild().

This removes a lot of complexity from the view_manager, while aligning
the v1 and v2 APIs a bit more.

SCN-914 #comment
Test: set_root_view shadertoy_client, tiles_ctl, image_grid_view, ran
system UI

Change-Id: I155448003e72ad51f5e94ab6815bbaf681b8e2ae