[debugger] Improve register and print output.
Updates the error message when a value is not available to differentiate
a variable being entirely optimized out vs. just not available at the
current address.

Changes register formatting. The new format removes the 32 bit grouping
for 64 bit registers because this makes it difficult to copy/paste the
hex values.

Adds decoding for small magnitude integers to also show their decimal
value. Removes left-0-padding. Removes some of the headings (name/value
is obvious in this context), and separate column handling for some
registers (they all use name/value/comment).

This changes the code factoring a bit. It moves away from using Err in
the register formatting. At this level it's not something the user could
have caused nor is it recoverable, so we should just do our best or
assert. Removes some of the duplicated code between the ARM and x86
formatting.

TEST=unit,manual

Change-Id: I9c29550d64bb2be533509c0168514ca92ddd9ff1