[fuzz] Save out to log
This CL changes Fuzzer::Start to set a "-jobs=1" option by default. This
doesn't change the default number of worker processes, but does save the
output to fuzz-0.log instead of sending it to stdout.  A side effect of
this change is that the 'wait_for_completion' argument is no longer
needed, as it can be inferred from the jobs option.

SEC-103

Tests: Added to the TestStart unit test
Change-Id: I273b75b5fbccd63ea19ec7924f5953e299159f6c