[services] Migrate listen.cmx to 'services'.
This requires some code changes. We recently changed the behavior of
ls /svc to accurately reflect the services present in the component/
environment (CP-124). `fx shell` runs ssh, which reaches the `listen`
daemon. The listen daemon spawns a process in its namespace. Right now
this is fine because the listen component uses
'deprecated-all-services'. But if we set a service whitelist, `ls /svc`
only sees a subset of system services (the services that are accessible
to `listen`.)

The solution is to spawn the process with chrealm.

CP-155 #done

TESTED=`fx shell 'ls /svc'`

Change-Id: I39f34d272ad80b34d9713458669987835007c92b