[docs] TLS documentation
This document explains many (but not all) details of how TLS works
and gives a few details of how our specific implementation works.
The primary goal is to first and foremost explain the TLS ABI
in detail. The second goal is to give a rough overview of how
TLS works in Zircon's implementation.

Test: This is a doc

Change-Id: I7f45a187fd4370fd0f8b2e2ebcb969bfd48655dc