[debugger] Step over unsymbolized functions by default.
Adds an option to control whether "step" steps over unsymbolized code or
not. Normally it steps over, but "step -f" will force stepping into
assembly if it is encountered.

Adds support for automatically stepping over function calls with no
symbols. This happens for libc calls in a few places.

Additional tests are added for both cases (stop and continue) for so
thunks and function calls with no symbols.

TEST=unit,manual

Change-Id: Id5888cf101ee93223cbf6efb97fdce8ed0e76419