[services] Remove |host_directory| from CreateNestedEnvironment.
All users of CreateNestedEnvironment pass empty |host_directory|. All
that's left is to remove it from the interface. This CL also removes
CreateNestedEnvironmentOptions and replaces it with
CreateNestedEnvironment that takes options.

There will be a companion cl in peridot for the migration in that
layer.

CP-124 #done

TESTED=run garnet tests

Change-Id: Ib7178cb57383566641daef9dbf29eb1c0c5bcb31