[virtcon] Don't update framebuffer with no display
This change reduces how much work the virtcon does when it doesn't have
ownership of the display. First, add checks for ownership to all checks
for the presence of a display (i.e. vc_gfx being non-null), so we treat
not having ownership the same as not having a framebuffer to render to.
Second, add ownership checks to all checks for vc->active, so we treat
not having ownership as all virtcons being inactive.

Test: Check that output during no display appears when focus is gained
Change-Id: Ib8ca77b0c9a104131bd37756878b30f7136c1b38