[dev][nand][aml-rawnand] Fix uncorrectable ECC handling.
1) Fix handling of uncorrectable ECC errors, For non-randomized
data, uncorrectable ECC errors result in an error propagated up
to the NAND protocol, which retries the read thrice. For randomized
data a further check is necessary (which was previously wrong - reversed).
2) Add verbose error logging to the code handling uncorrectable ECC errors.
3) Fix error handling for the case where we timeout waiting for a read
interrupt.
4) Populate the ECC strength from the Page0 read (the ECC strength is
used in the handling of ECC uncorrectable errors for the randomized
case - not applicable to us right now, but might be later).

Bug: ZX-2616.

Test: Use Ruchira's code to do SDIO reads at intervals of our choice
to generate bus loading, and use Ricardo's tester for data validation.

Change-Id: I1be58d291f872cb91018aabd3c0673caa109c507