[escher] Avoid excessive framebuffer/descriptor-set cycling.
Our goal is to get to a state where there is one escher::Frame
created for every rendered Frame.  We are not currently able
to do this, e.g. we create a separate frame of type kTransfer
for BatchGpuUploader, even though this type of CommandBuffer
never touches framebuffers/descriptor-sets.

This CL addresses the problem by checking the frame type
in Escher::NewFrame(), and only cycling the allocators
when appropriate.

ES-103 #comment

TEST: No more validtion errors in Escher/Linux waterfall2.

Change-Id: I80e867f6b60ab40e06f928e51e4035c2b005af73