[ledger] Rename YesNoUnknown to PagePredicateResult.
Now the possible results are YES, NO and PAGE_OPENED (instead of UNKNOWN).

NO was previously also used for the case of the page being open when the operation
checking the predicate was called. Now, PAGE_OPENED is returned instead.
In all other cases, UNKNOWN is replaced by PAGE_OPENED, and YES and NO remain
as they were.

Test: LedgerManagerTest.* tests still pass
Change-Id: I0fe4898903ab1bee6db1a5e69495b851ecd63d02