[wayland] Add Interface::REQUESTS and Interface::EVENTS.
Currently messages are converted from raw |Message| buffers into the
|Interface::Request| objects by using a generated function that does
the mapping in a completely transparent way to the dispatcher. This
makes it difficult to provide a simple way for bindings to interact
with the 'new_id' and 'object' arguments.

To solve this, we'll instead provide the dispatcher with the list of
|ArgKind|s for each opeartion an object can receive. We'll then convert
the |FromMessage| and |IntoMessage| traits into something more like
|FromArgSlice| and |IntoArgSlice|.

This change simply starts emitting the data structures describing the
message structure. Right now this is a list of |ArgKind| lists. The
outer list is indexed by message opcode and the inner list is the set
of arguments to that operation. The other changes to start using these
new data structures will come in a following change.

Example scanner output:

    pub struct WlDisplay;

    impl Interface for WlDisplay {
        const NAME: &'static str = "wl_display";
        const VERSION: u32 = 1;
        const REQUESTS: MessageGroupSpec = MessageGroupSpec(&[
            // sync
            MessageSpec(&[
                ArgKind::NewId,
            ]),
            // get_registry
            MessageSpec(&[
                ArgKind::NewId,
            ]),
        ]);
        const EVENTS: MessageGroupSpec = MessageGroupSpec(&[
            // error
            MessageSpec(&[
                ArgKind::Object,
                ArgKind::Uint,
                ArgKind::String,
            ]),
            // delete_id
            MessageSpec(&[
                ArgKind::Uint,
            ]),
        ]);
        type Request = WlDisplayRequest;
        type Event = WlDisplayEvent;
    }

Test: fx run-test fuchsia_wayland_core_lib_test_rustc
Change-Id: I1f9b9f1aab98984e3f0f1586ce32221dadd9f487