[cloud_provider] define a commit pack struct and add a library
This patch adds a concept of a serialized commit pack to the cloud
provider API, and a private library that we will use on both ends to
pack and unpack commits into a buffer. In a subsequent CL we will switch
to using commit packs everywhere that vector<Commit> is used today.

This is so that we don't hit the channel max message size limit when
retrieving long commit histories from the cloud.

LE-397
Test: unit-tests included

Change-Id: I6fd0940acb6c69d66bbc4550cffe1609e81f3e5b