[mediaplayer] Internal renderer rename
This CL is part 2 of a multi-stage sequence that renames the
primary FIDL interfaces for audio, from AudioOut and AudioIn
to the more intuitive AudioRenderer and AudioCapturer.

Just as its predecessor CL migrates the media directory, this CL
migrates the garnet/.../mediaplayer directory away from the
terms audio_out, audio_in, AudioOut and AudioIn, primarily in
function names, variables, parameters and comments. This CL
contains no changes to FIDL interfaces -- that will come in
a successor CL.

Test: build, CQ

Change-Id: I1ea7b47114beeb7c9ef404c3db859d7e943b382e