[mediaplayer] update and enable FidlDecoder
FidlDecoder was broken by the new allocation scheme and turned off.
This CL brings it up-to-date with respect to that scheme and enables
it for some video encodings.

https://fuchsia-review.googlesource.com/c/garnet/+/206577 must be
submitted before this CL. There's no dependency in terms of git, but
the video renderer won't support the NV12 output of FidlDecoder
without that change.

TESTED: manually tested on VIM2
Change-Id: Id7d2da33e170e2fb05fe81009362f442209893e6