[dev][nand][aml-rawnand] Handle (read) DMAs being aborted.
From the data we have, it seems clear that in the event of high data
bus contention, NAND read DMA operations get aborted. The data DMA is
aborted and the DMA of the read completion state is never done. This
means the driver cannot rely on the read completion state being accurate
or correct. Pre-initialize the read completion state with "bad" values,
So that if the DMA is aborted, and the read completion state is never
posted to memory, then we fail the read. The read is retried from the
NAND protocol layer.

Bug: ZX-2616.

Test: Run Ricardo's read verification test (with a higher number of
reads per block, running in a forever loop), with Ruchira's code to do
sdio reads in the background once every 100us. Verified that the DMAs
are aborted by pre-seeding the data buffer in the NAND controller driver
with a known value, and seeing that same value posted back to the
application in a "successful" read (before this change in place).

Change-Id: Ia0d27af53ce2535302eac6f9985bf7b4edc474fb