[debugger] Print reference values.
This is the first pass of handling references. It adds reference support
to the FormatValue code, with a special format (it shows the address
like a pointer, but also automatically dereferences the value unlike a
pointer). This format approximately matches other debuggers.

This does not yet handle references in expressions. For example, it will
not allow array[ref_value] where ref_value is a reference. That will
come in a later pass.

TEST=unit,manual

Change-Id: I5f8fdd81198df6ff3167e1ae2ea33a1977648cf9