[tracing] Implement support for virtual threads
The following API additions for the tracing system makes it
possible to create virtual threads that represent timelines
that execute non-CPU tasks. These virtual threads will be
visualized in the trace viewer as normal threads. Only major
difference is that duration events will only have a wall time.
CPU time is always zero.

Virtual threads are tied to the process from which they are
created and updated. Virtual threads are identified by an id
which the caller of the trace event macro controls. The
existing TRACE_NONCE() helper macro can be used to generate
a unique virtual thread id.

Each virtual thread is also associated with a name. The name
must be specified for each trace event macro call.

Test: trace-test
Change-Id: I0a5d343c0d025a56837c09fb65ea68efaaca27e8