[fdio] Prepare for fdio to defer to zxio
This CL factors the bulk of the machinery for fuchsia.io.Node-based file
descriptors into zxio. The fdio library still contains the glue for
binding these implementations to file descriptors.

This CL contains all the code for the transformation but the code is
disabled behind a build flag. A future CL will flip the flag.

Test: No behavior change
Change-Id: I61616b0dff9bee46a7e6871f4b94d424b8c4ac1f