[fidl][golang] Prepare bindings for new marshaling
For bindings to work with the new marshaling, we need:
- Expose a Marshaler method which returns a Marshaler. The expected way
is to CreateLazyMarshaler one in a variable, and then simply return this
pre-built lazy marshaler.
- Annotate structs with a tag containing size and alignement (on a _
struct{} field)
- Annotate unions with a tag containing size and alignement (on the
discriminator field)
- Annotate bounded fields (and/or those containing handles) with their
bounds.

Note that netstack_service_impl_test had to be modified because we now
use unexported fields on FIDL structs, and go-cmp barfs on these. Tried
with cmp.AllowUnexported and still got failures, so switching to
reflect.DeepEqual.

Test: go test fidl/compiler/backend/golang/ir and manual inspection of
bindings

Change-Id: I686bb3c188af1af6510a9119335b9ec4dd382561