[pci] Refactor kPCI ECAM Address Translation
Factor PCI ECAM address translation into an address provider class.
This allows the platform to decide which mechanism for address
translation it wants to provide.

This patch is only a refactoring for now and lays the groundwork
for allowing the DesignWare driver to provide its own address
translation mechanism which will be implemented in a following patch.

TEST: Built Zircon with `./scripts/buildall -crA`
Booted in Qemu x86|arm64 and booted on Pixelbook.

Change-Id: If56c9724c0c6510033a86ac11ecb7f604bbaaaf5