[debugger] Fix crash repeating previous line.
The Console saved the previous Command object to replay it if the user
presses enter with no input at the prompt. But the Command object can
not be cached since it contains references to the current frame, thread,
etc.

This manifested as a crash if the previuos context was not valid (for
example, if the program exited).

TEST=manual

Change-Id: I222c0f794da82c0e2618a1268ef014e70e138ac0