[fuzz] Add 'watch'.
This CL adds the 'watch' command, which will check the fuzzer
periodically and alert the user when it encounters a bug.  It also
improves how start and check collect and save output, and includes a new
'--output' option to specify where that information should be saved.

SEC-104

Tests: Manually tested.

Change-Id: I18947054c8388db26052036fad9c2d666d0e61d3