[packages] Add netcfg to netstack config so it shows up in SDK
netstack now requires the netcfg package to function (discover
interfaces) but the sdk configuration did not include this since it
included the netstack package directly instead of 'network' which also
contains an HTTP implementation. This moves the import to netstack so
it's always there.

Test: //scripts/packages/visualize_hierarchy.py \
--packages garnet/packages/sdk/garnet \
--output garnet_sdk.dot
dot -Tsvg garnet_sdk.dot -ogarnet_sdk.svg

Observe prod/netcfg in the hierarchy

Change-Id: Idfa406428f66f131481eef03780f62c29e443f46