[wayland][bridge] Add Display, ClientConnection.
This change integrates the wayland dispatcher and the generated library
for the wayland core protocol into the bridge. We spawn a new async task
for each new client connection that will poll for messages and dispatch
them in to the connections |ObjectMap|.

Much of the glue will need some work yet, but this gets us to the point
where the client can bind to the registry and the server will emit an
event for each global.

Test: fx run-test wayland_bridge_test_lib_test_rustc
Change-Id: Ic0b2e3fc9c738003adaf981f2cef9ceed546386b