[debugger] Fix crash in "finish"
Finish would assert and then crash if executed on the oldest stack
frame. It was not expecting the null IP or that the unwind controller
might be null (an artifact of trying to handle this case but doing so
incompletely).

Adds clarification to the "unwind" function about why we have a last
stack frame will a null IP. I went to fix this and realized why we need
it.

Refactors some tests that mock the RemoteAPI. The Finish test previuosly
used the "thread impl" test because the thread controller tests didn't
exist when it was written. This was moved over to being a thread
controller test.

Unified the RemoteAPI mocks for thread controller tests and thread impl
tests. This reduces the amount of test code.

Some refactoring of the finish thread controller test to make the new
test duplicate less code.

TEST=unit-manual

Change-Id: I0c988ef986d263d9a0647bea72c0185228a34d31