[Scenic] Add FIDL interface for ViewState
Adds a ViewState to the fidl gfx events, with an is_rendering field.
The ViewState is sent to the corresponding ViewHolder whenever it
changes. Is_rendering is set when the View has content that will be part
of the next Scenic frame, and can be used by the parent ViewHolder when
the child View has content that is presented.

Currently not hooked up the render pass.

Tested:
- client builds and runs.
- added unittests

Change-Id: I769d4ca1416e9a572d2937a1b09b83105581f228