[syslog] Fix invalid use of stack object as revealed by ASAN
Running scenic under ASAN revealed an invalid use of stack memory.
This was because the syslogger was building a list of tags to be
passed to the zircon logger by calling c_str() on each string during
a loop. That isn't valid. Instead I made the loop var be a reference.

Test: Run scenic under ASAN.
Change-Id: Ide6d8f27cc5bd39d5fd30811fafeef9e574c3392