[audio] Refactor Gain object
This CL refactors the Gain object into independent stages that
represent the front and back halves of the scaling applied during
a mix operation -- respectively called the source and dest gains.
During playback, these halves are the stream (renderer) gain and
the device (output) gain, respectively. During capture, the roles
are reversed: the source gain is the input device gain and the
dest gain is the stream (capturer) gain.

By incorporating a SetDestGain call, we can then add stateless
methods such as IsSilent() and IsUnity().  This also sets the
stage for the upcoming addition of IsRamping().

Finally, this CL also includes test coverage in audio_mixer_tests
specifically around SetDestGain(), the caching of these values,
and the IsUnity() and IsSilent() methods.

Test: additions to audio_mixer_tests (mentioned above), build, CQ

Change-Id: I86052848d301ce6515397a4e3ceae660a6978cd5