[images] exclude package files in meta/ from blobs
Ideally we'd re-wire the layout of the pm build process instead, so that
either it produces the correct manifest that we need, or it accepts separate
manifests for these cases.

For now this approach is the path of least resistance, with the long term
plan being to rewrite pm to better handle cases liek the above.

This reduces wasted space in blobfs and reduces the total number of blobs in
blobfs.

BLD-280 #done

Change-Id: I0b2e4efaa0fe2c0a03323c5ad7c01db62d082012