[wlan][mlme][ap] Use AID generated by SME
Since SME generates the AID now, it's no longer necessarily to do so in
the MLME. The BssClientMap is no longer necessary and is removed.
Additionally, it's no longer necessary to manually assign and release
the AID since the AID is owned by RemoteClient's AssociatedState, and
it's automatically "released" when RemoteClient moves to a different
state.

NET-1592 #done

Test: wlan_ap_unittest, manual sanity test connecting to unprotected AP
Change-Id: I24f20ed7f1e37b0391c3a89affb9461c7d904431