[wlan][minstrel] helper functions for comparison.
This is to fix a bug when using reference to a TxStats, the data is
overwritten by accident when the intention is to reassign the reference.

Extracting the logic into helper functions eliminates this problem and
made complicated logic more readable without clogging up the main
function.

Test: Build and manual smoke.
Also verified the correct behavior in local build.

Change-Id: I6707a77562d1c2f1f338a4d5c95e727ddade3ec3