[zircon][usb-request-cleanup]Remove bti_handle from usb-request
The bti_handle is used only during the physmap which is used only at the
usb infrastructure level and not by the client drivers. This change removes
the bti handle from usb_request_t, and also removes the need for it beeing needed
during the usb_request_alloc. In a later changeset, clients will be using
the usb_request_alloc directly from the usb-request library and not via
the protocol op. This changeset is in preparation for that to work.

Test: Full build. Netbooting on vim2. Plugged in a usb-mass-storage
device and verified lsusb works as expected.

Change-Id: I158917a9cfec1c3b0f2b9ec905c17dbfcc6ecbc7