[dw][pcie] PCI BDF Translation for DesignWare
A previous patch factored the responsibility of translating a BDF
address into a system address into the PcieAddressProvider interface.

This patch implements and exercises a DesignWare implementation of the
PcieAddressProvider.

This means that Zircon's PCIe bus driver is now aware of the PCI root
bridge as well as the downstream device.

Test: Booted QEMU x64 and observed devices using `k lspci`
Booted on the Pixelbook and observed the same.
Booted on Gauss and observed PCIe Root Bridge _and_ downstream device
listed with `k lspci`

Change-Id: Ie3879b38f314f2f2c47ce273343c88a7c7b91fef