[wayland] Wire up enum arguments.
This provides an Enum<E> argument to each RequestReceiver. The inner
'E' type will model all the values known statically (included in the
wayland protocol), but the wrapper type will also transport the
unmodeled values in the case the receiver is prepared to handle them.
Receviers can unwrap the Enum and return an error if they're not
prepared to handle unmodeled enum values:

Note we do not support unmodeled enum values for output events since
we should have full control over adding any new values to the model
as needed.

Ex:
  enum Request {
    enum_value: Enum<MyEnum>,
  }
  enum Event {
    enum_value: MyEnum,
  }
  enum MyEnum {
    Value1,
    Value2,
    Value3,
  }

Test: fx run-test wayland_scanner_back_end_test_lib_test_rustc
Change-Id: I26fd0dbf7f089591ef72c22679349596b80ca3f1