[socket] Use first datagram for available size
Change socket_read with null buffer and size to only return the first
datagram size instead of the size in all available datagrams.

This means that the *actual* is always equal to "the number of bytes you
need to allocate to read everything socket_read will return next".

Clarify the documentation language around socket_read with null buffer
and size.

Test: datagram_write_basic
Change-Id: I3438e6f961e5bde931578f4f3dd2a09a31552afb