[wlan][mlme] Bump auto-deauth to ~10s
This change makes the STA's association a little bit more robust
as the timeout for auto-deauthentication is longer.
This change does not explain why the STA sometimes doesn't receive
Beacons for 5 seconds and we'll have to do a proper investigation
of this issue to understand what's going on.

Test=Included

NET-1601 #comment

Change-Id: I5d970657feb4644af44dfb144b6279d4b405a86f