[wlan][mlme] Don't wait for a beacon when joining
We don't really seem to use the beacon in any way. And we already have
the necessary information in BSSDescription from a previous beacon. So
this step only introduces an extra delay to the connection process.
We can instead proceed directly to sending an auth frame and wait for a
response to that.

The real motivation for this change is QCA98xx not reporting beacons
when not associated.

TEST=Manually verified on QCA6174, QCA98xx and Ralink that we can still
successfully associate.

Change-Id: I716d95ba722af152360d2af841468624a3156ae6