[debugger] Allow multiple variable resolutions.
Add the ability for SymbolVariableResolver to support multiple pending
resolution requests. The resolver is associated with the frame, so if a
request is sent that returns asynchronously, it allows the possibility
of getting another request. I haven't seen this happen in practice.

Moves the per-request state to a reference-counted helper object that
the SymbolVariableResolver passes around throughout the request. Adds a
test that nested requests work properly.

TEST=unit,manual

Change-Id: Idb6e69bee52256124852ac48c86c920bbbbfa781